# Optimize ToI420 conversion  by using sync.Pool