--- a/pkg/io/video/convert.go
+++ b/pkg/io/video/convert.go
@@ -61,46 +61,24 @@ func imageToYCbCr(dst *image.YCbCr, src image.Image) {
 	}
 }
 
+// bytePool stores slices to be reused
+// New method is not set as the slice size
+// should be allocated according to subsample ratio
+var bytesPool = sync.Pool{}
+
 // ToI420 converts r to a new reader that will output images in I420 format
 func ToI420(r Reader) Reader {
 	var yuvImg image.YCbCr
 
-	// Uint8 slice pool created to reuse slices from previous calls
-	// New method is not set as the slice size should be allocated subsample ratio
-	// TODO: add reconfigurable pool limit
-	// TODO: make through gorouting
-	poolSize := make(chan struct{}, 1<<5)
-
-	uintPool := sync.Pool{}
-
-	putSlice := func(dst []uint8) {
-		// Check Limit of the pool
-		// If more than expected, do not put into the pool
-		select {
-		case poolSize <- struct{}{}:
-			// Put the slice into the pool
-			// There is no need to "clean" slice
-			uintPool.Put(dst)
-		default:
-		}
-	}
-
 	getSlice := func(cLen int) []uint8 {
 		// Retrieve slice from pool
-		// Decrement current pool size
-		var dst []uint8
+		dst, ok := bytesPool.Get().([]byte)
 
-		select {
-		case <-poolSize:
-			dst = uintPool.Get().([]uint8)
-		default:
-		}
-
-		// Compare capacity of retrieved slice
-		// If less than expected, reallocate new slice
-		if dst == nil || cap(dst) < 2*cLen {
+		// Compare value or capacity of retrieved object
+		// If less than expected, reallocate new object
+		if !ok || cap(dst) < 2*cLen {
 			// Allocating memory for Cb and Cr
-			dst = make([]uint8, 2*cLen, 2*cLen)
+			dst = make([]byte, 2*cLen, 2*cLen)
 		}
 
 		return dst
@@ -124,14 +102,14 @@ func ToI420(r Reader) Reader {
 			dst := getSlice(cLen)
 			yuvImg = i444ToI420(yuvImg, dst)
 			releaseFunc = func() {
-				putSlice(dst)
+				bytesPool.Put(dst)
 			}
 		case image.YCbCrSubsampleRatio422:
 			cLen := yuvImg.CStride * (yuvImg.Rect.Dy() / 2)
 			dst := getSlice(cLen)
 			yuvImg = i422ToI420(yuvImg, dst)
 			releaseFunc = func() {
-				putSlice(dst)
+				bytesPool.Put(dst)
 			}
 		default:
 			return nil, releaseFunc, fmt.Errorf("unsupported pixel format: %s", yuvImg.SubsampleRatio)