@neversi 2023-03-07
@edaniels, regarding to expanding pool to other frame decoders, I think it is possible, I will research it on weekends.